+++ assets/app.stories-CFscaRTg.js
@@ -1,0 +1,80 @@
+import{j as e}from"./iframe-CBGcgu9u.js";import{M as p}from"./index-r9jJnTS6.js";import{S as c}from"./index-DChAHxSX.js";import{A as t,M as d,C as l}from"./index-aAHwbgMY.js";import{O as v}from"./OverviewPage-JFtexNWk.js";import{H as u}from"./HomeIcon-DrFln7Ac.js";import"./FilledIcon-yTGXRcWU.js";import"./DropdownIcon-BaSW3IZg.js";import"./type-Bc9hys3F.js";import"./Heading-E73M5mBu.js";import"./AddIcon-Cw72PvZC.js";import"./useOutsideToggle--S2XFe4d.js";import"./DiviExperienceHelperIcon-D14cYLjQ.js";import"./PACLogo-BcQZnTDp.js";import"./index-DjmF8Zdq.js";import"./index-Vi1xmjYa.js";import"./index-BvFNopnz.js";import"./ChevronRightIcon-3oEZSop7.js";import"./BackIcon-DpFGefoA.js";import"./index-Dvfa5Z_Q.js";import"./index-DqLBRBSF.js";import"./index-CVIFn9np.js";import"./CloseIcon-DanjXDfh.js";import"./Text-jUIRRPpK.js";import"./index-gLfF79xc.js";import"./index-CQWP1Npk.js";import"./index-sEhXkRtb.js";import"./KeyIcon-DBGoID8j.js";import"./OpenIcon-Bc6mCY6P.js";import"./SupportIcon-CiCdQTXa.js";import"./UsersIcon-hXzHg0IR.js";import"./index-CHmk602f.js";import"./EyeIcon-B3IC4Trn.js";import"./index-Dq7h7Pqt.js";import"./index-DU-o1YNu.js";import"./InfoIcon-BjsuN96n.js";import"./StopIcon-nXdaKq3W.js";import"./TextLink-_d6ld3lE.js";const $={title:"PAC Universal Plugin UI/Layout/App",componnet:t,parameters:{pageLayout:"page"}},i=()=>{const r={avatar:"https://lh3.googleusercontent.com/ogw/AF2bZyg0NWhbUjvHEo5VTQIPmq2lgGyO4xDZeaWXPyHq5OknKQ=s32-c-mo",name:"Muhammad Touseef",greeting:"Good afternoon",pluginLogo:"https://dummyjson.com/image/74",pluginDescription:"A huge collection of incredibly handy features and tools to make your life easier when working on your Divi website - startup actions, general utilities, interface hacks, custom settings, global styles, learning resources, maintenance tools, code editors, accessibility improvements, alt text fixes, media hacks, local fonts, system report sharing, import/export tools, and so much more!",pluginName:"Divi Assistant",pacPluginNewsApi:"https://dummyjson.com/c/f371-88b4-4d1e-b88e",onPacCardClick:()=>{console.log("PAC Card Clicked")},docLink:"https://www.peeayecreative.com/docs/divi-assistant/",supportLink:"https://www.peeayecreative.com/support/",communityLink:"https://www.peeayecreative.com/community/space/divi-assistant/home"},s={version:"0.0.9",userAvatar:"https://bc3-production-assets-cdn.basecamp-static.com/4677743/people/BAhpBJQn1wI=--4b34d6ad2aa165f177c7e08ebb0fbb25e7fb5da0/avatar/in/260/90ad41c3cfd4e50976d4a0fad1ef44c8.avif?v=1",userName:"Muhammad Touseef",activeMenu:"",plugins:[{id:"divi-assistant",name:"Divi Assistant"},{id:"divi-dynamic-helper",name:"Divi Dynamic Content Helper"},{id:"divi-contact-form-helper",name:"Divi Contact Form Helper"},{id:"divi-custom-fields-helper",name:"Divi Custom Fields Helper"},{id:"divi-event-calendar-module",name:"Divi Events Calendar"},{id:"divi-taxonomy-helper",name:"Divi Taxonomy Helper"},{id:"divi-search-helper",name:"Divi Search Helper"},{id:"divi-table-of-contents-maker",name:"Divi Table of Contents Maker"},{id:"divi-carousel-maker",name:"Divi Carousel Maker"},{id:"divi-responsive-helper",name:"Divi Responsive Helper"},{id:"divi-social-sharing-buttons",name:"Divi Social Sharing Buttons Maker"},{id:"divi-timer-pro",name:"Divi Timer Pro"},{id:"divi-tabs-maker",name:"Divi Tabs Maker"}],activePluginId:"divi-responsive-helper",onPluginSelect:()=>{},onAvatarClick:()=>console.log("Go to Admin Page")},m=[{id:"overview",label:"Overview",icon:u}];return e.jsxs(t,{height:"400px",children:[e.jsx(p,{...s}),e.jsxs(d,{children:[e.jsx(c,{menuData:m,activeMenu:"overview",onActiveSubmenuChange:()=>{},onActiveMenuChange:()=>{},showAddNewMenu:!0,showAllPluginMenu:!0}),e.jsx(l,{children:e.jsx(v,{...r})})]})]})};i.__docgenInfo={description:"",methods:[],displayName:"App"};var n,a,o;i.parameters={...i.parameters,docs:{...(n=i.parameters)==null?void 0:n.docs,source:{originalSource:`() => {
+  const OverviewPageProps = {
+    avatar: "https://lh3.googleusercontent.com/ogw/AF2bZyg0NWhbUjvHEo5VTQIPmq2lgGyO4xDZeaWXPyHq5OknKQ=s32-c-mo",
+    name: "Muhammad Touseef",
+    greeting: "Good afternoon",
+    pluginLogo: "https://dummyjson.com/image/74",
+    pluginDescription: "A huge collection of incredibly handy features and tools to make your life easier when working on your Divi website - startup actions, general utilities, interface hacks, custom settings, global styles, learning resources, maintenance tools, code editors, accessibility improvements, alt text fixes, media hacks, local fonts, system report sharing, import/export tools, and so much more!",
+    pluginName: "Divi Assistant",
+    pacPluginNewsApi: "https://dummyjson.com/c/f371-88b4-4d1e-b88e",
+    onPacCardClick: () => {
+      console.log("PAC Card Clicked");
+    },
+    docLink: "https://www.peeayecreative.com/docs/divi-assistant/",
+    supportLink: "https://www.peeayecreative.com/support/",
+    communityLink: "https://www.peeayecreative.com/community/space/divi-assistant/home"
+  };
+  const MainheaderProps = {
+    version: "0.0.9",
+    userAvatar: "https://bc3-production-assets-cdn.basecamp-static.com/4677743/people/BAhpBJQn1wI=--4b34d6ad2aa165f177c7e08ebb0fbb25e7fb5da0/avatar/in/260/90ad41c3cfd4e50976d4a0fad1ef44c8.avif?v=1",
+    userName: "Muhammad Touseef",
+    activeMenu: "",
+    plugins: [{
+      id: "divi-assistant",
+      name: "Divi Assistant"
+    }, {
+      id: "divi-dynamic-helper",
+      name: "Divi Dynamic Content Helper"
+    }, {
+      id: "divi-contact-form-helper",
+      name: "Divi Contact Form Helper"
+    }, {
+      id: "divi-custom-fields-helper",
+      name: "Divi Custom Fields Helper"
+    }, {
+      id: "divi-event-calendar-module",
+      name: "Divi Events Calendar"
+    }, {
+      id: "divi-taxonomy-helper",
+      name: "Divi Taxonomy Helper"
+    }, {
+      id: "divi-search-helper",
+      name: "Divi Search Helper"
+    }, {
+      id: "divi-table-of-contents-maker",
+      name: "Divi Table of Contents Maker"
+    }, {
+      id: "divi-carousel-maker",
+      name: "Divi Carousel Maker"
+    }, {
+      id: "divi-responsive-helper",
+      name: "Divi Responsive Helper"
+    }, {
+      id: "divi-social-sharing-buttons",
+      name: "Divi Social Sharing Buttons Maker"
+    }, {
+      id: "divi-timer-pro",
+      name: "Divi Timer Pro"
+    }, {
+      id: "divi-tabs-maker",
+      name: "Divi Tabs Maker"
+    }],
+    activePluginId: "divi-responsive-helper",
+    onPluginSelect: () => {},
+    onAvatarClick: () => console.log("Go to Admin Page")
+  };
+  const sidemenuData = [{
+    id: "overview",
+    label: "Overview",
+    icon: HomeIcon
+  }];
+  return <AppLayout height="400px">
+      <MainHeader {...MainheaderProps} />
+      <MainLayout>
+        <Sidebar menuData={sidemenuData} activeMenu={"overview"} onActiveSubmenuChange={() => {}} onActiveMenuChange={() => {}} showAddNewMenu showAllPluginMenu />
+        <ContentLayout>
+          <OverviewPage {...OverviewPageProps} />
+        </ContentLayout>
+      </MainLayout>
+    </AppLayout>;
+}`,...(o=(a=i.parameters)==null?void 0:a.docs)==null?void 0:o.source}}};const ee=["App"];export{i as App,ee as __namedExportsOrder,$ as default};